Update patch set 6
Patch Set 6: Code-Review+1

This LGTM but I'm not the owner.

There is one unresolved question: whether "cgpt repair" works with off-device GPT storage. The test commands are usually having this form:

  $CGPT show $MTD ...

Your commands do not include the "$MTD" part.

If it works when you add that, please do so. But don't mind it if it fails the tests. It can be solved in the future.

Patch-set: 6
Reviewer: Gerrit User 1109105 <1109105@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1