vbnv: Do not clear fastboot_unlock if dev-mode is not active
The flag "fastboot_unlock_in_fw" decides whether device unlock is allowed in
firmware. This flag is not related to dev_boot_* flags. If this flag
is cleared every time dev mode is not active, then it will not be
possible to unlock the device in firmware at all.

BUG=None
BRANCH=None
TEST=Compiles successfully and "fastboot oem unlock" works in firmware
with fastboot_unlock_in_fw flag set in VbNvStorage.

Change-Id: I57ca21f335f9e6ead7997f2b13504402827aba08
Signed-off-by: Furquan Shaikh <furquan@google.com>