Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

Overall this looks pretty good.

I'd prefer that we implement a 2-bit dev_default_boot field, rather than a 1-bit dev_default_boot_legacy field.  This would also allow defaulting dev boot to USB, a feature which has been requested in the past.

So, 
00 = disk (default)
01 = usb (if dev_boot_usb=1, else boots disk)
10 = legacy (if dev_boot_legacy=1, else boots disk)
11 = (reserved for future use; for now, boots disk)

If we have the 2-bit field, then it's clearer that each of the single-bit fields is enabling a particular boot source, and the 2-bit field is setting the preferred source (if and only if it's enabled).

If we don't do this now, and we later add dev_default_boot_usb, then we'll have a very confusing set of flags.  Because what should happen in the following cases:
- dev_boot_usb = 1
- dev_boot_legacy = 1
- dev_default_boot_usb = 1
- dev_default_boot_legacy = 1

- dev_boot_usb = 1
- dev_boot_legacy = 0
- dev_default_boot_usb = 0
- dev_default_boot_legacy = 1

- dev_boot_usb = 0
- dev_boot_legacy = 0
- dev_default_boot_usb = 1
- dev_default_boot_legacy = 1

Which one gets priority?  And it also leads to the weirdness of mixing the default boot source with which boot sources are actually allowed/enabled, as you've seen in the patches so far.

Or it leads to a funky workaround in crossystem, where setting dev_default_boot_legacy=1 forces us also to set dev_boot_legacy=1 and dev_default_boot_usb=0.  (That is, we make the default boot fields radio buttons, where crossystem only lets you select one of them.)

If we make dev_boot_default a 2-bit field to start with, we never run into that problem.

Patch-set: 4
Label: Code-Review=-1