vboot_api_kernel: Power down if low power usage is requested
If we find that low power usage is requested by calling the new
VbExEcIsLimitPowerRequested() API, we may not have enough power to
continue booting. Wait for the low power condition to clear, and power
down if the condition is not cleared within 3 seconds.

BUG=chromium:537269
TEST=Manual on Glados. Set CHG_MW thresh to 20000, BAT_PCT to 50. Verify
that LIMIT_POWER host event is set until Zinger negotiates to 20V. Also
verify that we do not proceed with boot when Donette is plugged.
BRANCH=None
CQ-DEPEND=CL:307885

Change-Id: I77e6000aa8a44e3aca4fb5982e5b5f5191774989
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>