cgpt repair: update primary header when we move alternate
The primary GPT header has a link to the secondary, via the "alternate
LBA" field. So if we're updating the LBA of the secondary header, we
need to update the alternate LBA field on the primary too.

Noticed when testing GPU parted partition table checks after running
"cgpt repair" on a Chrome OS USB test image. cgpt thinks it repairs the
disk properly, but GNU parted complains about the mislink.

BRANCH=none
BUG=none
TEST=flash image to USB; run "cgpt repair" on host to expand to USB disk
     size; run "parted /dev/sdX check ...", look for GPT header errors;
     also boot samus from USB disk

Change-Id: I384a329e15ef6c3713d2d2613f2ed473cd35710c
Signed-off-by: Brian Norris <briannorris@chromium.org>