Update patch set 2
Patch Set 2:

(6 comments)

I have fixed all the comments on patch set 2 except the one for xxd. I will look to see if i can find a standard utility that can be used instead of xxd.

On the complication aspect as commented by vapier@, Nvidia mentioned that they do not want to include the pkc signing stuff in their cbootimage. Instead they expect us to use openssl to perform all the signing, which makes the script look really complicated.

One solution that we have here is to continue with the existing solution i.e. privately shared version of cbootimage which does all the work for us and keeps the script clean. In the future, if we need similar signing on any other Nvidia platform we can push for providing a generic script that handles all required platforms. 

Thoughts? If we are fine with this, I can abandon the current CL and punt the bug to future platform.

Patch-set: 2