vboot: Disable VBNV_OPROM_NEEDED after successful update
The VBOOT_OPROM_NEEDED flag is used for EC software sync when the
VBSD_EC_SLOW_UPDATE flag is set.

After a successful EC software sync vboot requests a reboot to disable
graphics but it is not clearing the VBNV flag first.  With vboot1 this
was getting cleared as a side effect of calling VbInit in normal mode.

BUG=chrome-os-partner:49560
BRANCH=glados
TEST=Enable EC_SLOW_UPDATE on chell and test EC software sync in normal
mode and ensure that it reboots and does not do graphics init if the
update is successful.

Change-Id: I2aa0c4c3b1ad357a5b8ddc14539e264a1f5b76b2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>