Update patch set 1
Patch Set 1:

> I think you might be misunderstanding the circumstances.

Yeah, okay, I did. I thought your goal was to avoid the resume time delay of running the whole verification chain from the start again (with all the long RSA operations). If you want to run the whole thing anyway and just compare that you got the exact same image as last time, this makes more sense. (I'm still a little worried about the digest size, though... maybe we could have vboot headers define a VB2_MAX_DIGEST_SIZE to at least abstract it out a little, so we run less of a risk of forgetting to update the size of the buffer passed in there if we ever end up implementing a longer digest type in the future.)

Some more thoughts on what you're doing, according to how I understand it now:

1. If you're running chromeos-firmwareupdate --mode=autoupdate, it will immediately change fw_try_next to the other slot. If you then suspend/resume and run the normal vb2api_fw_phase2(), it will select the wrong slot. You will need to add code to force it to choose the fw_prev_tried slot on S3 resume instead. (In fact, running the normal vboot API may have other unintended side effects like decrementing the try counter and things like that... we have to very carefully review which steps in it still make sense for resume.)
2. Are we going to have Skylake Chromeboxes? Where are you going to store the hash there?
3. If we didn't care about malicious attacks, this could be way simpler (I think the only relevant time the updater overwrites the current slot right now is during updates for non-WPed devices... and I think there's not really a reason to do that either, we could just change it). Is there really some security benefit of preventing this against an attacker who already had root access (to write the flash)... especially if all he could do is roll forward to a newer official release?
4. If there's really a security issue with that, another solution might be to store the booted firmware+key rollback version in the TPM and compare it on S3 resume. We'd just need to remember to always roll it forward when we're releasing a firmware update that could be exploited in this way (I assume we're mostly concerned about memory init updates?).

Patch-set: 1