Add battery-cutoff into NV requests.
DRAFT.

Add a new crossystem value "battery_cutoff_request" to indicate that
next reboot should cut-off battery and shutdown during firmware stages.

This request is primarily for factories to ship devices in an safe
state. Previously we do same thing by running "ectool battery-cutoff"
but that creates a problem that "ectool" (and the one to request for
cut-off) must live in developer mode, but the device must be shipped in
normal mode. The mode transition this can be solved by setting
"disable_dev_request=1", but that flag is also stored in NV storage and
may get lost once the system enters "battery cut-off" stage.

From the experience from Ryu, such settings (dev mode transition and
battery cut-off) must be done together inside firmware execution so we
have to create a new flag, battery_cutoff_request, for it.

BUG=none
TEST=none

Change-Id: I73ccae15b337cd65786106646546c67c155b8fa6