vboot: do not re-draw recovery screen unless necessary
When device is in recovery more and there is no kernel to boot, or
there is a kernel, but it is not valid, the device keeps polling
sources of the recovery image. Each poll cycle involves re-drawing the
screen, but it is not necessary, as the screen contents are not
supposed to change unless situation changes.

Let' re-draw the screen only when condition changes.

BRANCH=none
BUG=none
TEST=verified that drawing the screen is invoked only once while
     sitting in recovery.

Change-Id: I5c5ea925ab54be75b217e0e89493cd27f4b8799a
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>