Update patch set 2
Patch Set 2:

Dug into this a bit.

Take a look at VbDisplayScreenLegacy().  That's what VbDisplayScreen() used to do, and it skipped re-displaying screens.

It looks like when we moved to the new VbExDisplayScreen() code that broke.  That saves the currently displayed screen, but no longer has this check:

/* If requested screen is the same as the current one, we're done. */
        if (disp_current_screen == screen && 0 == force)
                return VBERROR_SUCCESS;

I think maybe we should move that check down to VbDisplayScreen() instead of this change.

Patch-set: 2