Update patch set 2
Patch Set 2:

Note for posterity: after reading through crbug.com/210352, IIUC, the decision to remove /boot from signed images was made in the past so that the AU payload size would be smaller.  This patch adds an exception based on the bios type, which is inferred from the kernel commandline.  As a result, the size of AU payloads will be larger for EFI based systems.

The original bug also mentions the possibility of populating EFI partition based on the kernel partition rather than /boot of rootfs.  I think this idea is worth revisiting because Lakitu board is using either BIOS or EFI in production.  If it is feasible, we can revert this patch.

Patch-set: 2