Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

Nits: Mixed tabs and spaces.

One question: The previous commit message says that either primary or secondary partitions can be marked IGNOREME. But this command does not support that second case. And we also don't test for that case in the previous commit either. So, the question is, do we want to support that case?

Patch-set: 1
Label: Code-Review=+2