Update patch set 2
Patch Set 2: Code-Review-2

> Patch Set 2: Commit message was updated.

VbDisplayScreenLegacy is used only for boards which store bitmap images in GBB. This is deprecated and supported only in firmware branches created for the legacy boards. So, if this patch is intended to fix the referenced bug, it should be fixed differently. That is, the image.bin needs to store bitmaps in cbfs not in GBB. This should work if bitmap_in_cbfs switch is USE'd (and the code around it hasn't been broken).

Patch-set: 2
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2