Update patch set 1
Patch Set 1:

(1 comment)

Note that this is a cherry-pick from ToT.  There was reformatting done in ToT between the Candy branch point and now, which explains the whitespace differences.

If we're going to make more substantial changes than whitespace, we should make them in ToT first (which would still require picking this change into the branch, followed by a cleanup change in ToT subsequently picked into the branch).

Patch-set: 1