Update patch set 1
Patch Set 1:

I made vboot_host build with the correct set of tlcl functions for tpm2. But the existing support is rudimentary. It builds, but tpmc, mount-encrypted and similar still need to be updated (and/or the supported set of tlcl functions should be extended for tpm2) in order to actually work. Not tested atm - just verified that it builds in both modes.

Patch-set: 1