Update patch set 2
Patch Set 2:

OK, I can simplify this one then to always use AUTHREAD in NVRead in Tlcl - doesn't matter if in depthcharge or userland. 

Are we sure that there are no indexes that must be created with only PPREAD set (i.e. non-readable from userland but readable from fw)? If I simplify the implementation, depthcharge will also switch to always using AUTHREAD in NVRead: this Tlcl routine is universal - it is used for reading any index we need to access in fw or from utilities like tpmc.

Am I right that if we have only AUTHREAD set in DefineSpace, and somebody tries to read from the index using TPM_RH_PLATFORM, the read will fail, even though an empty password is provided?

If so, dropping PPREAD from DefineSpace (in Tlcl and other places like coreboot) is to be a separate CL since it needs to be ensured that other places in fw don't rely on PPREAD set. And, at least the similarly-purposed marshal_nv_read() in coreboot uses TPM_RH_PLATFORM when reading from NVRam.

Patch-set: 2