Update patch set 2
Patch Set 2:

It's just for NVRam reads, and only because (and if) we always set AUTHREAD for NVRam indexes we care about reading from Tlcl. Writes still require proper platform authorization etc.
Of course, a full-blown API should allow specifying the authorization from the top level when the code supposedly knows where it is called from and what needs to be done. But then it won't be a lightweight Tlcl API anymore.

Patch-set: 2