Update patch set 3
Patch Set 3:

FWIW, I think this is a bug in the TPM implementation. From the spec:

"The attributes of the Index determine what authorizations are allowed. TPMA_NV_PPREAD allows the Index to be read using Platform Authorization; TPMA_NV_OWNERREAD allows the Index to be read using Owner Authorization; TPMA_NV_AUTHREAD allows the Index to be read using the authValue of the Index; and TPMA_NV_POLICYREAD allows the Index to be read if the authPolicy of the Index is satisfied."

Patch-set: 3